+++ packages/kofono/src/validator/isValid/IsNotValidValidator.spec.ts
@@ -8,7 +8,7 @@ describe("isNotValidValidator", () => {
     beforeAll(async () => {
         form = await K.form({
             propA: K.string().$v(v => v.notEmpty()),
-            propB: K.string().$q(v => v.notValid("propA")),
+            propB: K.string().$q(v => v.isNotValid("propA")),
         });
     });